Functional navigation to genre after uploading Gcal, and able to update Firebase with the busy timings
